+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+- **Added:** `itron::processor::Processor` now implements `TryFrom<usize>`
+
 ## [0.1.4] - 2021-06-30
 
 - **Changed:** `itron::task::Builder::{stack → stack_size}`